fins-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
supabase==2.0.2
httpx>=0.24.0,<0.25.0
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
xgboost==2.0.2
prophet==1.1.4
python-dateutil==2.8.2
pytz==2023.3
python-dotenv==1.0.0
alembic==1.13.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9 